Split tests into 2 files
